Assignments

+++ src/main.cpp
@@ -10,10 +10,10 @@ int yyparse();
 struct Options
 {
     bool parse_tree = false;
+    bool keep_cpp = false;
 };
 
-int
-main(int argc, char **argv)
+int main(int argc, char **argv)
 {
     Options options;
     for (int i = 0; i < argc; i++)
@@ -22,6 +22,10 @@ main(int argc, char **argv)
         {
             options.parse_tree = true;
         }
+        else if (std::string(argv[i]) == "--keep-cpp")
+        {
+            options.keep_cpp = true;
+        }
     }
 
     // Parsing
@@ -29,14 +33,13 @@ main(int argc, char **argv)
     root = simplify_tree(root);
 
     if (options.parse_tree)
-    {
         print_node(root);
-    }
 
     // Generate code
     generate_code(root, "out.cpp");
     compile_code("out.cpp", "program");
-    system("rm out.cpp");
+    if (!options.keep_cpp)
+        system("rm out.cpp");
 
     return 0;
 }
